requirements.txt
flask
langchain
langchain-openai
langchain-community
openai
faiss-cpu
python-dotenv
pypdf
tiktoken